+++ configurationfiles/cgitest.conf
@@ -40,6 +40,7 @@ server {
 	location /cgi/ {
 		abspath /www/cgi;
 		allow_methods GET POST;
+		cgi_methods GET POST;
 		cgiexecutable /usr/bin/python3;
 		cgi_extension .py;
 		upload_path /www/uploads/;
