+++ expected.txt
@@ -1,0 +1,5 @@
+HTTP/1.1 403 Forbidden
+Content-Length: 139
+Content-Type: text/html; charset=UTF-8
+
+<html><head><title>403 Forbidden</title></head><body><h1>403 Forbidden</h1><p>The server encountered an error: Forbidden.</p></body></html>
