--- home/images/random.file
@@ -1,1 +1,0 @@
-Temporary file content.
